Make stacks updateable
When we configure servers (via OS::Nova::Server) for boot-from-volume,
and we actually want to use a single persistent volume to boot from,
then this creates a catch-22 when we update the server resource in
such a way that Heat detects that the server requires replacement:
Heat then tries to spin up a new server before removing the old one,
but at that stage it obviously cannot attach the boot volume.

One way in which we can circumvent this problem is to use BFV, but to
not actually boot from a single boot volume, but rather to recreate
the boot volume each time the server is recreated.

Thus:

* Drop the my_dockerhost_boot_volume resource.
* Instead, create the boot volume directly from the
  block_device_mapping_v2 map in the OS::Nova::Server resource.