Add a persistent volume for use by Docker
In order to ensure that Docker has something to work with that
persists across server replacements, add a (non-boot) persistent
volume via an OS::Cinder::Volume resource, and manage it via
OS::Cinder::VolumeAttachment.

Then, use cloud-init to manage this volume as a btrfs filesystem, and
mount it to /var/lib/docker.