Add a WaitCondition
Without a WaitCondition resource, Heat considers the creation of the
stack (and any subsequent stack updates) complete as soon as the
OS::Nova::Server instance has booted. This is not usually helpful;
instead we want things to be considered "done" when the stack has been
fully configured.

The most straightforward way to do that in a Heat context is to add a
WaitCondition, and invoke it at the end of the cloud-init sequence via
runcmd.